Creates delegation of DCIntrospect and update the demo.
Allows user to extends useful functionality, like keyboard inputs, for their own use cases.
As the project is iOS 4 compatible, delegate property is set as assign, which requires the user to nil the delegate upon deallocation.
It would be better if we can make use of the weak property attribute instead which release the user the responsibility of setting the delegate to nil.